carte statique
