+++ src/routes/+layout.ts
@@ -1,2 +1,2 @@
 export const prerender = true;
-export const csr = true;
+export const csr = false;
